implement editable & undoable tempo and time signature objects
Care has been taken so that all changes to the tempo map happen while the
engine is paused.

Also added some useful utilities:
- ranges::all_equal
- commands::ChangeQObjectPropertyCommand
- actions::QObjectPropertyOperator

Implements #5183.

+++ src/structure/arrangement/arranger_object_factory.h
@@ -264,7 +264,7 @@ public:
         {
           if constexpr (std::is_same_v<ObjT, AutomationPoint>)
             {
-              obj->setValue (*automatable_value_);
+              obj->setValue (static_cast<float> (*automatable_value_));
             }
         }
